This commit introduces a series of improvements to the application's user experience on desktop devices, based on user feedback.
Key changes include:

1.  **Desktop Layout:** The main application container on desktop is now constrained to a 9:16 aspect ratio, creating a centered, phone-like viewing experience.
2.  **PWA Installation Flow:**
    - The PWA installation prompt is now correctly positioned within the new desktop container.
    - The phone icon in the top bar has been removed.
    - The 'Install' button in the prompt now opens a modal with a QR code for mobile installation, as direct PWA installation is disabled for desktop.
3.  **UI Fixes:**
    - The z-index of the login panel has been adjusted to ensure it appears correctly under the top bar.
    - The login button's text color has been fixed to remain white during interaction, removing the previous pink hover/focus state.
