+++ app/layout.tsx
@@ -29,7 +29,6 @@ export default function RootLayout({
       <body className={cn("antialiased", inter.className)}>
         <Providers>
           <AppLayout>{children}</AppLayout>
-          <PWAInstallPrompt />
         </Providers>
         <Script
           data-name="BMC-Widget"
