+++ components/DesktopLayout.tsx
@@ -8,6 +8,7 @@ import AccountPanel from './AccountPanel';
 import { AnimatePresence } from 'framer-motion';
 import TopBar from './TopBar';
 import BottomBar from './BottomBar';
+import PWAInstallPrompt from './PWAInstallPrompt';
 
 export default function DesktopLayout({ children }: { children: React.ReactNode }) {
     const context = useVideoGrid();
@@ -22,14 +23,15 @@ export default function DesktopLayout({ children }: { children: React.ReactNode 
 
     return (
         <div
-            className="relative h-full w-full flex justify-center"
+            className="relative h-screen w-full flex justify-center items-center bg-zinc-900"
         >
-            <div className="relative h-full w-full max-w-sm overflow-hidden border-x border-white/10 flex flex-col bg-black">
+            <div className="relative h-full max-h-[960px] aspect-[9/16] overflow-hidden rounded-md shadow-lg border-x border-white/10 flex flex-col bg-black">
                 <TopBar />
                 <main className="flex-1 overflow-y-auto">
                     {children}
                 </main>
                 {activeVideoRef && <BottomBar videoRef={activeVideoRef} isActive={true} />}
+                <PWAInstallPrompt />
             </div>
 
             {/* Conditionally render modals based on the centralized state */}
